+++ src/main/java/com/nylas/JsonHelper.java
@@ -31,7 +31,7 @@ public class JsonHelper {
 				.add(new Integration.IntegrationListCustomAdapter())
 				.add(new Grant.GrantCustomAdapter())
 				.add(new Grant.GrantListCustomAdapter())
-				.add(new UASLoginInfo.UASLoginInfoCustomAdapter())
+				.add(new LoginInfo.LoginInfoCustomAdapter())
 				.add(Date.class, new Rfc3339DateJsonAdapter().nullSafe())
 				.build();
 	}