+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This section contains changes that have been committed but not yet released.
 ### Added
 
 - Added support for Delta
-- Added UAS Support (Integrations API, Grants API, UAS Hosted Authentication)
+- Added support for new (beta) Integrations authentication (Integrations API, Grants API, Hosted Authentication for Integrations)
 
 ### Changed
 
@@ -16,6 +16,8 @@ This section contains changes that have been committed but not yet released.
 ### Deprecated
 
 ### Fixed
+
+- Fixed incorrect property name for `Event.Notification.minutes_before_event`
 
 ### Removed